Rename header and sidebar components in edit-site
