+++ packages/edit-site/src/components/sidebar-edit-mode/navigation-menu-sidebar/style.scss
