+++ packages/edit-site/src/components/sidebar-edit-mode/style.scss
@@ -1,4 +1,4 @@
-.edit-site-sidebar {
+.edit-site-sidebar-edit-mode {
 	width: $sidebar-width;
 
 	> .components-panel {
