+++ packages/edit-site/src/components/sidebar-edit-mode/template-card/style.scss
