+++ packages/edit-site/src/components/sidebar-edit-mode/template-card/template-areas.js
