+++ packages/edit-site/src/style.scss
@@ -3,16 +3,16 @@
 @import "./components/block-editor/style.scss";
 @import "./components/code-editor/style.scss";
 @import "./components/global-styles/style.scss";
-@import "./components/header/style.scss";
-@import "./components/header/document-actions/style.scss";
+@import "./components/header-edit-mode/style.scss";
+@import "./components/header-edit-mode/document-actions/style.scss";
 @import "./components/navigation-sidebar/navigation-toggle/style.scss";
 @import "./components/navigation-sidebar/navigation-panel/style.scss";
 @import "./components/list/style.scss";
 @import "./components/add-new-template/style.scss";
-@import "./components/sidebar/style.scss";
-@import "./components/sidebar/settings-header/style.scss";
-@import "./components/sidebar/navigation-menu-sidebar/style.scss";
-@import "./components/sidebar/template-card/style.scss";
+@import "./components/sidebar-edit-mode/style.scss";
+@import "./components/sidebar-edit-mode/settings-header/style.scss";
+@import "./components/sidebar-edit-mode/navigation-menu-sidebar/style.scss";
+@import "./components/sidebar-edit-mode/template-card/style.scss";
 @import "./components/editor/style.scss";
 @import "./components/template-details/style.scss";
 @import "./components/create-template-part-modal/style.scss";
